# packages/kytchen-rlm/kytchen_rlm-1.0.0.tar.gz/kytchen_rlm-1.0.0/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kytchen Cloud API",
    "description": "Kytchen Cloud API - BYOLLM context processing with sandboxed execution.\n\n## Features\n- **Datasets (Pantry)**: Upload and manage datasets for context processing\n- **Query**: Run natural language queries against datasets using your LLM\n- **Tools (Prep)**: Execute sandboxed Python code, search, and peek context\n- **Evidence (Sauce)**: Track and export evidence from analysis runs\n\n## Authentication\nAll endpoints (except /healthz) require Bearer token authentication:\n`Authorization: Bearer kyt_sk_...`\n\nGet your API key at https://kytchen.dev",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/kytchens": {
      "get": {
        "tags": [
          "kytchens"
        ],
        "summary": "List Kytchens",
        "description": "List all Kytchens in the workspace.",
        "operationId": "list_kytchens_v1_kytchens_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KytchenListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "kytchens"
        ],
        "summary": "Create Kytchen",
        "description": "Create a new Kytchen.",
        "operationId": "create_kytchen_v1_kytchens_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KytchenCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KytchenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/kytchens/{kytchen_id}/git/status": {
      "get": {
        "tags": [
          "git"
        ],
        "summary": "Git Status",
        "description": "Show repository status.",
        "operationId": "git_status_v1_kytchens__kytchen_id__git_status_get",
        "parameters": [
          {
            "name": "kytchen_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kytchen Id"
            }
          },
          {
            "name": "short",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Use short format",
              "default": false,
              "title": "Short"
            },
            "description": "Use short format"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitCommandResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/kytchens/{kytchen_id}/git/diff": {
      "get": {
        "tags": [
          "git"
        ],
        "summary": "Git Diff",
        "description": "Show diff between commits/files.",
        "operationId": "git_diff_v1_kytchens__kytchen_id__git_diff_get",
        "parameters": [
          {
            "name": "kytchen_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kytchen Id"
            }
          },
          {
            "name": "target",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Commit/branch to diff against",
              "title": "Target"
            },
            "description": "Commit/branch to diff against"
          },
          {
            "name": "staged",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Show staged changes",
              "default": false,
              "title": "Staged"
            },
            "description": "Show staged changes"
          },
          {
            "name": "stat",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Show diffstat",
              "default": false,
              "title": "Stat"
            },
            "description": "Show diffstat"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitCommandResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/kytchens/{kytchen_id}/git/log": {
      "get": {
        "tags": [
          "git"
        ],
        "summary": "Git Log",
        "description": "Show commit history.",
        "operationId": "git_log_v1_kytchens__kytchen_id__git_log_get",
        "parameters": [
          {
            "name": "kytchen_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kytchen Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Number of commits to show",
              "default": 10,
              "title": "Limit"
            },
            "description": "Number of commits to show"
          },
          {
            "name": "oneline",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Use one-line format",
              "default": false,
              "title": "Oneline"
            },
            "description": "Use one-line format"
          },
          {
            "name": "author",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by author",
              "title": "Author"
            },
            "description": "Filter by author"
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Show commits since date",
              "title": "Since"
            },
            "description": "Show commits since date"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitCommandResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/kytchens/{kytchen_id}/git/blame": {
      "get": {
        "tags": [
          "git"
        ],
        "summary": "Git Blame",
        "description": "Show per-line commit info for a file.",
        "operationId": "git_blame_v1_kytchens__kytchen_id__git_blame_get",
        "parameters": [
          {
            "name": "kytchen_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kytchen Id"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "File path to blame",
              "title": "Path"
            },
            "description": "File path to blame"
          },
          {
            "name": "line_start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start line",
              "title": "Line Start"
            },
            "description": "Start line"
          },
          {
            "name": "line_end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End line",
              "title": "Line End"
            },
            "description": "End line"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitCommandResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/kytchens/{kytchen_id}/git/show": {
      "get": {
        "tags": [
          "git"
        ],
        "summary": "Git Show",
        "description": "Show contents of a specific commit.",
        "operationId": "git_show_v1_kytchens__kytchen_id__git_show_get",
        "parameters": [
          {
            "name": "kytchen_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kytchen Id"
            }
          },
          {
            "name": "ref",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Commit reference to show",
              "default": "HEAD",
              "title": "Ref"
            },
            "description": "Commit reference to show"
          },
          {
            "name": "stat",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Show diffstat",
              "default": false,
              "title": "Stat"
            },
            "description": "Show diffstat"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitCommandResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/kytchens/{kytchen_id}/git/branch": {
      "get": {
        "tags": [
          "git"
        ],
        "summary": "Git Branch",
        "description": "List branches.",
        "operationId": "git_branch_v1_kytchens__kytchen_id__git_branch_get",
        "parameters": [
          {
            "name": "kytchen_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kytchen Id"
            }
          },
          {
            "name": "all",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Show all branches",
              "default": false,
              "title": "All"
            },
            "description": "Show all branches"
          },
          {
            "name": "verbose",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Show last commit",
              "default": false,
              "title": "Verbose"
            },
            "description": "Show last commit"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitCommandResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GitCommandResult": {
        "properties": {
          "stdout": {
            "type": "string",
            "title": "Stdout"
          },
          "stderr": {
            "type": "string",
            "title": "Stderr"
          },
          "exit_code": {
            "type": "integer",
            "title": "Exit Code"
          }
        },
        "type": "object",
        "required": [
          "stdout",
          "stderr",
          "exit_code"
        ],
        "title": "GitCommandResult",
        "description": "Result of a git command."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KytchenCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "Kytchen name"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1,
                "pattern": "^[a-z0-9-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug",
            "description": "URL-friendly slug (auto-generated if not provided)"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Kytchen description"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "private",
              "unlisted"
            ],
            "title": "Visibility",
            "description": "Kytchen visibility",
            "default": "private"
          },
          "fork_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fork From",
            "description": "Kytchen ID to fork from"
          },
          "recipe_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipe Id",
            "description": "Recipe ID to create from"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "KytchenCreate"
      },
      "KytchenListResponse": {
        "properties": {
          "kytchens": {
            "items": {
              "$ref": "#/components/schemas/KytchenResponse"
            },
            "type": "array",
            "title": "Kytchens",
            "description": "List of kytchens"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "Total count"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "More results available",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "kytchens",
          "total"
        ],
        "title": "KytchenListResponse"
      },
...
      "KytchenResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Kytchen ID"
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "URL-friendly slug"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Kytchen name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Kytchen description"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "private",
              "unlisted"
            ],
            "title": "Visibility",
            "description": "Visibility"
          },
          "forked_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forked From",
            "description": "Kytchen ID this was forked from"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Last update timestamp"
          },
          "dataset_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Dataset Count",
            "description": "Number of datasets in pantry",
            "default": 0
          },
          "total_size_bytes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Size Bytes",
            "description": "Total pantry size",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "visibility",
          "created_at",
          "updated_at"
        ],
        "title": "KytchenResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}